MFB-1676: extract the packed standalone programs into white_labels/
The 17 calculators still living in <wl>/pe/{member,spm,tax}.py that belong to no
family: CO's andcs/oap/fatc/care-worker credit, MA's eaedc/mbta/ssp, TX's
ccs/dart/harris_rides, KS cdcc, IL ibccp, MO pts/wftc, NC scca, and the federal
acp/pell_grant.

Each now has its own directory with room for a spec and tests.
