+++ programs/framework/tests/test_receipt_contract.py
@@ -16,7 +16,6 @@ Three things can silently break it:
 
 from django.test import TestCase
 
-from programs.programs.co.pe.member import AidToTheNeedyAndDisabled
 from programs.programs.federal.pe.spm import SNAP_BASE_INPUTS
 from programs.framework.pe_dependencies import member, receipt_contract, spm
 from programs.programs.cross_white_label.medicaid.base import Medicaid
@@ -31,6 +30,7 @@ from programs.programs.cross_white_label.head_start.base import HeadStart
 from programs.programs.cross_white_label.early_head_start.base import EarlyHeadStart
 from programs.programs.cross_white_label.liheap.tx import TxCeap
 from programs.programs.cross_white_label.medicaid.disability.il_aabd import IlAabd
+from programs.programs.white_labels.co.andcs.calculator import AidToTheNeedyAndDisabled
 
 # The amount inputs predate the contract; everything else arrived with it.
 UNGATED_FIELDS = {"ssi", "tanf"}
