+++ programs/programs/co/pe/member.py
@@ -4,53 +4,3 @@ from screener.models import HouseholdMember
 from programs.programs.cross_white_label.medicaid.base import Medicaid
 from programs.programs.cross_white_label.wic.base import Wic
 from programs.programs.cross_white_label.csfp.base import CommoditySupplementalFoodProgram
-
-
-class AidToTheNeedyAndDisabled(PolicyEngineMembersCalculator):
-    program_code = "andcs"
-    pe_name = "co_state_supplement"
-    pe_inputs = [
-        dependency.member.SsiCountableResourcesDependency,
-        # co_state_supplement tops up SSI, so it reads the `ssi` amount the receipt
-        # contract supplies (reported where reported, simulated-and-suppressed otherwise).
-        *dependency.receipt_contract,
-        dependency.member.IsBlindDependency,
-        dependency.member.IsDisabledDependency,
-        dependency.member.SsiEarnedIncomeDependency,
-        dependency.member.SsiUnearnedIncomeDependency,
-        dependency.member.AgeDependency,
-        dependency.member.TaxUnitSpouseDependency,
-        dependency.member.TaxUnitHeadDependency,
-        dependency.member.TaxUnitDependentDependency,
-        dependency.household.CoStateCodeDependency,
-    ]
-    pe_outputs = [dependency.member.Andcs]
-
-
-class OldAgePension(PolicyEngineMembersCalculator):
-    program_code = "oap"
-    pe_name = "co_oap"
-    pe_inputs = [
-        dependency.member.SsiCountableResourcesDependency,
-        dependency.member.SsiEarnedIncomeDependency,
-        dependency.member.SsiUnearnedIncomeDependency,
-        dependency.member.AgeDependency,
-        dependency.member.TaxUnitSpouseDependency,
-        dependency.member.TaxUnitHeadDependency,
-        dependency.member.TaxUnitDependentDependency,
-        dependency.household.CoStateCodeDependency,
-    ]
-    pe_outputs = [dependency.member.Oap]
-
-
-class FamilyAffordabilityTaxCredit(PolicyEngineMembersCalculator):
-    program_code = "fatc"
-    pe_name = "co_family_affordability_credit"
-    pe_inputs = [
-        dependency.member.AgeDependency,
-        dependency.member.TaxUnitDependentDependency,
-        dependency.household.CoStateCodeDependency,
-        dependency.member.TaxUnitSpouseDependency,
-        *dependency.irs_gross_income,
-    ]
-    pe_outputs = [dependency.member.FamilyAffordabilityTaxCredit]
