+++ programs/programs/ks/pe/tax.py
@@ -3,22 +3,3 @@ import programs.framework.pe_dependencies as dependency
 from programs.programs.cross_white_label.cdcc.base import Cdcc
 from programs.programs.cross_white_label.eitc.base import Eitc
 from programs.programs.cross_white_label.ctc.base import Ctc
-
-
-class KsCdcc(PolicyEngineTaxUnitCalulator):
-    program_code = "ks_cdcc"
-    pe_name = "ks_cdcc"
-    pe_inputs = [
-        dependency.member.AgeDependency,
-        dependency.member.TaxUnitHeadDependency,
-        dependency.member.TaxUnitSpouseDependency,
-        dependency.member.TaxUnitDependentDependency,
-        dependency.member.IsDisabledDependency,
-        dependency.member.IsIncapableOfSelfCareDependency,
-        dependency.member.FullTimeCollegeStudentDependency,
-        dependency.spm.ChildCareDependency,
-        dependency.member.CareExpensesDependency,
-        dependency.household.KsStateCodeDependency,
-        *dependency.irs_gross_income,
-    ]
-    pe_outputs = [dependency.tax.KsCdcc]
