+++ programs/programs/ma/pe/spm.py
@@ -2,27 +2,3 @@ from programs.framework.pe_base import PolicyEngineSpmCalulator
 import programs.framework.pe_dependencies as dependency
 from programs.programs.cross_white_label.snap.base import Snap
 from programs.programs.cross_white_label.ssi.base import Ssi
-
-
-class MaEaedc(PolicyEngineSpmCalulator):
-    program_code = "ma_eaedc"
-    pe_name = "ma_eaedc"
-    pe_inputs = [
-        dependency.spm.MaEaedcLivingArangementDependency,
-        dependency.spm.CashAssetsDependency,
-        dependency.spm.PreSubsidyChildcareExpensesDependency,
-        dependency.spm.MaEaedcNonFinancialCriteria,
-        dependency.member.EmploymentIncomeDependency,
-        dependency.member.SelfEmploymentIncomeDependency,
-        dependency.member.InvestmentIncomeDependency,
-        dependency.member.PensionIncomeDependency,
-        dependency.member.SocialSecurityIncomeDependency,
-        dependency.member.AgeDependency,
-        dependency.member.TaxUnitHeadDependency,
-        dependency.member.TaxUnitSpouseDependency,
-        dependency.member.TaxUnitDependentDependency,
-        dependency.member.MaTotalHoursWorkedDependency,
-        dependency.member.IsDisabledDependency,
-        *Ssi.pe_inputs,
-    ]
-    pe_outputs = [dependency.spm.MaEaedc]
