+++ programs/programs/ma/pe/tests/test_mbta.py
@@ -4,8 +4,7 @@ Unit tests for MaMbta MBTA service area filtering.
 
 from django.test import TestCase
 from unittest.mock import Mock, MagicMock
-
-from programs.programs.ma.pe.member import MaMbta
+from programs.programs.white_labels.ma.mbta.calculator import MaMbta
 
 
 class TestMaMbta(TestCase):
