+++ programs/programs/white_labels/federal/pell_grant/calculator.py
@@ -1,0 +1,21 @@
+"""PellGrant."""
+
+from programs.framework.pe_base import PolicyEngineMembersCalculator
+import programs.framework.pe_dependencies as dependency
+
+
+class PellGrant(PolicyEngineMembersCalculator):
+    program_code = "pell_grant"
+    pe_name = "pell_grant"
+    pe_inputs = [
+        dependency.member.PellGrantDependentAvailableIncomeDependency,
+        dependency.member.PellGrantCountableAssetsDependency,
+        dependency.member.CostOfAttendingCollegeDependency,
+        dependency.member.PellGrantMonthsInSchoolDependency,
+        dependency.tax.PellGrantPrimaryIncomeDependency,
+        dependency.tax.PellGrantDependentsInCollegeDependency,
+        dependency.member.TaxUnitDependentDependency,
+        dependency.member.TaxUnitHeadDependency,
+        dependency.member.TaxUnitSpouseDependency,
+    ]
+    pe_outputs = [dependency.member.PellGrant]
