+++ programs/programs/white_labels/ma/eaedc/calculator.py
@@ -1,0 +1,29 @@
+"""MaEaedc."""
+
+from programs.framework.pe_base import PolicyEngineSpmCalulator
+from programs.programs.cross_white_label.ssi.base import Ssi
+import programs.framework.pe_dependencies as dependency
+
+
+class MaEaedc(PolicyEngineSpmCalulator):
+    program_code = "ma_eaedc"
+    pe_name = "ma_eaedc"
+    pe_inputs = [
+        dependency.spm.MaEaedcLivingArangementDependency,
+        dependency.spm.CashAssetsDependency,
+        dependency.spm.PreSubsidyChildcareExpensesDependency,
+        dependency.spm.MaEaedcNonFinancialCriteria,
+        dependency.member.EmploymentIncomeDependency,
+        dependency.member.SelfEmploymentIncomeDependency,
+        dependency.member.InvestmentIncomeDependency,
+        dependency.member.PensionIncomeDependency,
+        dependency.member.SocialSecurityIncomeDependency,
+        dependency.member.AgeDependency,
+        dependency.member.TaxUnitHeadDependency,
+        dependency.member.TaxUnitSpouseDependency,
+        dependency.member.TaxUnitDependentDependency,
+        dependency.member.MaTotalHoursWorkedDependency,
+        dependency.member.IsDisabledDependency,
+        *Ssi.pe_inputs,
+    ]
+    pe_outputs = [dependency.spm.MaEaedc]
